sb: implement DMA timeout, large enough for KryptEgg to still work. Waiting for DMA forever makes a lockup to some progs. :(
dsp: fix fifo timing calculation wrt pcm_frame_period_us() again, and
verified with duke3d

+++ src/base/dev/sb16/sb16.h
@@ -75,7 +75,7 @@ struct sb_struct {
   uint8_t  mixer_regs[256];
   uint8_t  mixer_index;
   uint8_t  E2Count;
-  int      busy;
+  uint8_t  busy;
 /* All values are imperical! */
 #define SB_DSP_CMD_BUF_SZ 8
   uint8_t  command[SB_DSP_CMD_BUF_SZ];
